+++ next/src/components/ui/icons/CloseSmallIcon.tsx
@@ -1,0 +1,21 @@
+/**
+ * Icon of a small cross which should be used when something can be cancelled by
+ * clicking on it.
+ */
+export default function CloseSmallIcon({
+  height = '24px',
+  width = '24px',
+  ...props
+}: Readonly<React.ComponentProps<'svg'>>) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      height={height}
+      viewBox="0 -960 960 960"
+      width={width}
+      {...props}
+    >
+      <path d="m336-280-56-56 144-144-144-143 56-56 144 144 143-144 56 56-144 143 144 144-56 56-143-144-144 144Z" />
+    </svg>
+  )
+}