+++ next/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - migrate to latest oidc-spa (v10)
 - add `getAccessToken` & `decodedIdToken` to params of legacy App
+- series are no longer mandatory in questionnaire details
 
 ## [2.5.0](https://github.com/InseeFr/Pogues/releases/tag/2.5.0) - 2026-03-06